machine-learning/pyproject.toml
[tool.poetry]
name = "machine-learning"
version = "0.1.0"
description = ""
authors = ["Bruce  <97542494+BruceNiemi@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
gliner = "^0.2.15"
ollama = "^0.4.7"
pydantic-settings = "^2.8.1"
extract-msg = "^0.53.2"
pillow = "^11.1.0"
pypdf2 = "^3.0.1"
docx = "^0.2.4"
python-multipart = "^0.0.20"
python-docx = "^1.1.2"
pytesseract = "^0.3.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
